"Wrap hyprpolkitagent removal in conditional check to avoid errors if package is not installed in secure-boot script"

+++ .local/share/scripts/arch/systemd-boot/secure-boot/hyprland.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-pacman -Rsc hyprpolkitagent --noconfirm # with Replace polkit-kde-authentication-agent-1
+if pacman -Qi hyprpolkitagent &>/dev/null; then
+    pacman -Rsc hyprpolkitagent --noconfirm # with Replace polkit-kde-authentication-agent-1
+fi
 
 PACMAN_PACKAGES+=('hyprland' 'kwalletmanager' 'kwallet-pam' 'sddm' 'sddm-kcm' 'polkit' 'xorg-xwayland'
     'xorg-xeyes' 'xorg-xlsclients' 'xdg-desktop-portal-hyprland' 'qt5-wayland' 'qt6-wayland' 'hyprpaper'
